util: move node-only encounter printing to its own file
This makes it easier to include encounter_tools in a web page.
